backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9
python-dotenv==1.0.1

# Background jobs (optional)
celery==5.4.0
redis==5.0.8

# Validation/serialization
marshmallow==3.21.3
marshmallow-sqlalchemy==1.1.0

# HTTP
requests==2.32.3

# QR Codes
qrcode==7.4.2
Pillow==10.4.0

# Storage (optional S3)
boto3==1.34.144
